refactor: move useResizeObserver into its own file
